Add models folder
